fix(program): bool_t is one byte, not four
C defines the type as one byte -- typeDefinitions.h:10, `typedef bool bool_t`.
Four program owners aliased it to u32:

  zig_src/shell/program/{lbl_gto_xeq,next_step,clcvar,decode}.zig

so every store through that alias was four bytes wide. lbl_gto_xeq.zig has two:
`pemCursorIsZerothStep = 1` in fnReturn and `hourGlassIconEnabled = 1`. Both
globals are defined bool_t -- one byte -- in c47.zig, so each store wrote
01 00 00 00 and put three zero bytes past the end of a one-byte flag.

Caught on the Windows lane with a data breakpoint that resolved the faulting
instruction and its target:

  store is in  fnReturn+0x1c0
  victim  is   pemCursorIsZerothStep+0x0
  code: c7 05 f9 4f 08 01 01 00 00 00      MOV DWORD PTR [rip+disp], 1

Under ELF the three stray bytes land in alignment padding after the flag, so
nothing is observable and no test can see it. Under COFF each export gets its own
.bss$name COMDAT and the linker reorders them; there it had placed plotSelection
one byte after the flag and plotStatMx three bytes after, so fnReturn cleared
plotStatMx[0]. Every stat plot then lost its matrix name and reported error 28,
"No statistic data present", with the statistics themselves intact -- the six
graphs_cov failures that had the Windows lane red since the 4697e526a resync.

check-c-type-alias-widths.sh is added and wired into the guard job so no owner
can disagree with another about a C type's width again. The class is invisible in
isolation: the defining owner is right, the declaring owner is self-consistent,
and only the combination is wrong. Verified against this defect -- restoring
`const bool_t = u32;` makes the gate fail.

+++ zig_src/shell/program/clcvar.zig
@@ -21,7 +21,7 @@ const extra_info: bool = frontier_build_options.extra_info_on_calc_error;
 // Types
 // ---------------------------------------------------------------------------
 const calcRegister_t = i16;
-const bool_t = u32;
+const bool_t = u8; // C bool_t is one byte (typeDefinitions.h: typedef bool bool_t); u32 here made every store through it write 3 bytes past the global
 const angularMode_t = c_int;
 
 const abi = @import("abi"); // shared ABI bindings
