+++ zig_src/shell/program/decode.zig
@@ -34,7 +34,7 @@ else
 // Types
 // ---------------------------------------------------------------------------
 const calcRegister_t = i16;
-const bool_t = u32;
+const bool_t = u8; // C bool_t is one byte (typeDefinitions.h: typedef bool bool_t); u32 here made every store through it write 3 bytes past the global
 const angularMode_t = c_int;
 
 const real34_t = abi.Real34;
